Add a mechanism for categorizing errors
This change adds a custom error struct which contains an error
code to categorize the error. This allows clients to determine
whether or not an error is a system error or caused by user
mistake (such as operating on a cluster which does not exist).
This is particularly helpful for REST clients which need to
report proper HTTP status codes.